Reduce/centralize gitignored files
This update will stop gitignoring a bunch of files that contain keys & such. Don't worry: everything we need to hide but one instance of firebase-config is in a single folder, at the root, called secret. These files should not change as we dev, so download this folder from teams once, put it in the root of this repo, and dont worry about security again.