Changed res.render to res.redirect
Basically, after receiving a callback from the API we were rendering the chat page at the url the server listens at for a response from Spotify. That's why it kept throwing errors and ultimately redirecting to index. Now it redirects to the chat.html url, but seems to reinitialize after two messages.